feat: resolve some conflicts (#5)
There was a typo in the switch_case code which made the grammar
ambiguous. The other two fixes were due to specifying a higher
precedence for the inline expressions when used with the switch case
branches. I verified this behavior and the inline expressions take
precedence over the preceding inline keyword in parsing.